second commit

+++ src/main/java/cs/mum/edu/orangeteam/compro/Repository/TMRecordRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TMRecordRepository extends JpaRepository<TMRecord,Integer> {
+public interface TMRecordRepository extends JpaRepository<TMRecord,Long> {
 }
